amibroker-zerodha-trading/requirements.txt
eventlet==0.36.1
Flask==3.0.3
Flask_SocketIO==5.3.6
kiteconnect==5.0.1
psycopg2==2.9.9
psycopg2_binary==2.9.9
python-dotenv==1.0.1
python-socketio==5.11.3
websockets==12.0
